+++ CHANGES.rst
@@ -22,6 +22,7 @@ Improvements
 - Pre-fill the registration form picture field with the user's custom profile
   picture when available (:pr:`7447`, thanks :user:`moliholy, unconventionaldotdev`)
 - Allow users to mark contributions as favorites (:issue:`3524`, :pr:`7256`)
+- Allow choosing whether to clone registration tags (:pr:`7506`)
 
 Bugfixes
 ^^^^^^^^